Controller User delete , create

+++ Back/app/Controller/deleteUser.php
@@ -2,10 +2,10 @@
 
 $id = isset($_POST["id"]) ? $_POST["id"] : "";
 
-if(!empty($id)){
+ $userdelete = false;
+if(!empty($id) && (\app\core\User::isAdmin($_SESSION['id'], $_SESSION['login']) )){
   
    $user = new \app\core\User();
    $user->deleteUser($id);
-   
-    
+   $userdelete = true;
 }
